Postgres connection, EF migrations, seed data

+++ Models/MenuItem.cs
@@ -2,10 +2,9 @@
 {
     public class MenuItem
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public required string Name { get; set; }
         public string? Description { get; set; }
-
         public decimal? Price { get; set; }
     }
 }
